Add files via upload
Part1：In this project, the screen receives acoustic signals generated by the vibration of trees and converts them into visual signals.  So when the audience touches the tree trunk, the straight lines on the screen will transform into wavy lines of varying curvature.  Different sound-producing locations correspond to different positions on the screen, so when vibration occurs, wavy lines in different positions on the screen will move.  I'm interested in the correlation between force and visual effects, as well as the systematic movement of the wavy lines.
Part2：This code simulates grid structure of nodes and connecting lines. Drag mouse to cut the connecting wires, creating a widening gap in the grid. Parameters such as grid count, friction, and force multipliers are set. Using node and link classes define the behavior and appearance of nodes and links, including methods for updating their positions and displaying them on the canvas.Using the nodes are evenly distributed on the canvas, and links are established between nearby nodes.Drag mouse to cut the link, which triggers the mouseDragged() function.Using the undo() function, the cut can be undone when U or space is pressed.